# Speed up configure_i18n by precomputing paths and indexing descendants