--- a/rake/tasks/cldr.rake
+++ b/rake/tasks/cldr.rake
@@ -50,6 +50,30 @@ namespace :cldr do
     end
 
     desc <<~DESCRIPTION
+      Generate indices of files in the CLDR data
+
+      This task generates a list of all files in the CLDR data, which is used to speed up the loading of the data.
+
+      eg.: bundle exec rake cldr:data:generate_paths
+    DESCRIPTION
+    task :generate_paths do
+      cldr_locale_paths = Dir[File.join(Worldwide::Paths::CLDR_ROOT, "locales", "*", "*.{yml,rb}")].map do |path|
+        path.delete_prefix(Worldwide::Paths::GEM_ROOT)
+      end
+      File.write(Worldwide::Paths::CLDR_LOCALE_PATHS_FILE, cldr_locale_paths.join("\n"))
+
+      other_data_paths = Dir[File.join(Worldwide::Paths::OTHER_DATA_ROOT, "*", "*.{yml,rb}")].map do |path|
+        path.delete_prefix(Worldwide::Paths::GEM_ROOT)
+      end
+      File.write(Worldwide::Paths::OTHER_DATA_PATHS_FILE, other_data_paths.join("\n"))
+
+      region_data_paths = Dir[File.join(Worldwide::Paths::REGIONS_ROOT, "*", "*.{yml}")].map do |path|
+        path.delete_prefix(Worldwide::Paths::GEM_ROOT)
+      end
+      File.write(Worldwide::Paths::REGION_DATA_PATHS_FILE, region_data_paths.join("\n"))
+    end
+
+    desc <<~DESCRIPTION
       Generate additional data from the CLDR data, including missing plurals.
 
       Run this after importing and patching data from CLDR.