… data/cldr/paths.txt
… data/other/paths.txt
… data/regions/paths.txt