### What are you trying to accomplish?

In our codebase, we perform `Worldwide.configure_i18n` on every boot, which means every time we run tests. 
Recent versions of Worldwide have decreased the time required from ~350ms down to <200ms, but it's still quite a bit.

Here's a profile of `configure_i18n` taken before this PR:

<img width="1146" alt="image" src="https://github.com/user-attachments/assets/4eb7776c-668b-49c9-9d1b-bf22d7195019" />

[Profile before](https://github.com/user-attachments/files/18098893/vernier-before.json)

We can see that 2 things take up most of the time: 
- `Dir[]` (`cldr_locale_paths` is clearly visible, 2 others appear nested in `other_data_path` and `regions_data_path`)
- `expanded_locales_from_configuration` via `Array#include?`

Given the files in `data` are static, and the output of `Dir[]` is known at gem-package time, we can use this knowledge to pre-generate files that act as indices to the yaml files.

Second, instead of scanning each locale within the `fallback`, we can make a reverse-index and use that to figure out the descendants. 

<img width="790" alt="image" src="https://github.com/user-attachments/assets/fbd8e2f0-2d38-4db4-9f67-859370fe4c23" />

With these 2 changes, `configure_i18n` goes from 178ms to ~10ms on my machine.

There are other improvements we could make, like performing the matches only on the index (don't need to `match` the `GEM_ROOT` prefix), but I'm happy with this, for now.

### What approach did you choose and why?

Path of least resistance

### What should reviewers focus on?

1. Is there a better place for the rake task? I don't know anything about this gem
2. Are the changes _correct_ ?

### Checklist

* [ ] I have added a CHANGELOG entry for this change (or determined that it isn't needed)